Update and some clean change

+++ Ap.Demo.Infrastructure/Migrations/MyCitiesContextModelSnapshot.cs
@@ -37,8 +37,8 @@ namespace Ap.Demo.Infrastructure.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<int>("Population")
-                        .HasColumnType("int");
+                    b.Property<long>("Population")
+                        .HasColumnType("long");
 
                     b.HasKey("Id");
 
@@ -52,28 +52,28 @@ namespace Ap.Demo.Infrastructure.Migrations
                             Id = 1,
                             CountryId = 1,
                             Name = "Antwerpen",
-                            Population = 530000
+                            Population = 530000L
                         },
                         new
                         {
                             Id = 2,
                             CountryId = 1,
                             Name = "Gent",
-                            Population = 260000
+                            Population = 260000L
                         },
                         new
                         {
                             Id = 3,
                             CountryId = 2,
                             Name = "Amsterdam",
-                            Population = 820000
+                            Population = 820000L
                         },
                         new
                         {
                             Id = 4,
                             CountryId = 3,
                             Name = "Parijs",
-                            Population = 2140000
+                            Population = 2140000L
                         });
                 });
